delete custom_loss_model
