+++ pixyz/models/__init__.py
@@ -1,13 +1,13 @@
+from .model import Model
 from .vae import VAE
 from .vi import VI
 from .ml import ML
 from .gan import GAN
-from .custom_loss_model import CustomLossModel
 
 __all__ = [
+    'Model',
     'ML',
     'VAE',
     'VI',
     'GAN',
-    'CustomLossModel',
 ]
